test(e2e): dogfood playwright-rust main — Locator::drag_to migration
Pin slp-e2e + theoria-e2e to playwright-rust `main` (the v0.15.0-dev
line) and migrate every drag gesture off the manual
`mouse.down` → button-held `mouse.move_to` → `mouse.up` pattern, which
hangs on headless Linux under the new driver (Chromium 149). The fix is
`Locator::drag_to`, wrapped in three shared helpers in tests/common:

- `drag_to_ft(source, yard, ppf, fx, fy)` — drag an element (object body,
  resize/edge handle, node) onto a yard point in world feet.
- `drag_within_ft(yard, ppf, sfx, sfy, tfx, tfy)` — canvas-to-canvas drag
  in feet (no source element).
- `drag_within_px(yard, sx, sy, tx, ty)` — canvas drag by exact
  yard-relative pixels, for gestures where the page-pixel path matters
  (grid-snap assertions).

Also adapts to the new f64 `Mouse` coordinate API (drops `as i32` casts
on the surviving hover-only moves in placement_preview) and bumps the CI
browser install to Playwright 1.61.1.

Net -107 lines; the whole slp-e2e suite (57 tests) is green against the
new driver locally.

+++ crates/slp-e2e/tests/placement_preview.rs
@@ -48,7 +48,7 @@ async fn the_preview_ghost_tracks_the_pointer_and_matches_the_armed_shape() -> R
     // shape preview.
     let mouse = page.mouse();
     mouse
-        .move_to((x + 30.0 * ppf) as i32, (y + 15.0 * ppf) as i32, None)
+        .move_to(x + 30.0 * ppf, y + 15.0 * ppf, None)
         .await
         .context("hover with nothing armed")?;
     expect(page.locator("[data-testid='yard'] .placement-object-preview").await)
@@ -59,7 +59,7 @@ async fn the_preview_ghost_tracks_the_pointer_and_matches_the_armed_shape() -> R
     // Arm the (round) fire pit; hovering now shows a circular preview.
     arm_object(&page, "fire-pit").await?;
     mouse
-        .move_to((x + 32.0 * ppf) as i32, (y + 15.0 * ppf) as i32, None)
+        .move_to(x + 32.0 * ppf, y + 15.0 * ppf, None)
         .await
         .context("hover the yard with the fire pit armed")?;
     let preview = page
@@ -77,7 +77,7 @@ async fn the_preview_ghost_tracks_the_pointer_and_matches_the_armed_shape() -> R
 
     // Move elsewhere: the same preview circle follows.
     mouse
-        .move_to((x + 45.0 * ppf) as i32, (y + 8.0 * ppf) as i32, None)
+        .move_to(x + 45.0 * ppf, y + 8.0 * ppf, None)
         .await
         .context("hover a different point")?;
     let cx2 = preview
